chore: moved sync from table component to the helpers
