+++ libs/docs/platform/table/examples/platform-table-tree-example.component.ts
@@ -388,6 +388,7 @@ const ITEMS: ExampleItem[] = [
                                     },
                                     {
                                         name: 'Beam Breaker B-2',
+                                        selected: true,
                                         category: 'other',
                                         description: 'sapien in sapien iaculis congue',
                                         price: {
